Fix 64-bit warnings in libcchash

+++ shotgun/external_libs/libcchash/hashtable.c
@@ -21,8 +21,8 @@ static const unsigned int primes[] = {
 50331653, 100663319, 201326611, 402653189,
 805306457, 1610612741
 };
-const unsigned int prime_table_length = sizeof(primes)/sizeof(primes[0]);
-const float max_load_factor = 0.65;
+const unsigned int prime_table_length = (unsigned int)(sizeof(primes)/sizeof(primes[0]));
+const float max_load_factor = 0.65F;
 
 /*****************************************************************************/
 struct hashtable *
